# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenAPI natation project backend documentation",
        "description": "backend is used by smartphone application and desktop front",
        "contact": {
            "email": "**"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/athlete/{id}/programmes": {
            "get": {
                "tags": [
                    "Athletes"
                ],
                "summary": "Get athlete programs for mobile interface",
                "description": "Récupère les programmes d'un athlète pour interface mobile (annexe 9-1-B)",
                "operationId": "getAthletePrograms",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Athlete ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Athlete programs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "programs": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}